Remove redundant migration screen. (#101862)
* Redirect from the site credentials step directly to /sites/overview

This is the first step towards removing the SITE_MIGRATION_ASSISTED_MIGRATION step.

* Redirect from the site authorization step directly to /sites/overview

This is the first step towards removing the SITE_MIGRATION_ASSISTED_MIGRATION step.

* Fixing site-migration tests after removing assisted migration step

* Remove references to ASSISTED_MIGRATION step

We'll delete the actual files in a separate PR

* Copy updates on the migration-started-difm

* Remove unneeded query args when redirecting to the overview

* Remove unneeded if block

* Remove old TODO

* Remove unused type property

* Ensure we have a valid siteId before resetting the store

* Update so the 'Let us guide you' link creates the ZD ticket

Previously clicking that link would send you to the importer-migrate-message step which would create the ticket. Now that step is gone so we need to handle it from the credentials step (before we redirect to /overview)

* Rename ticket creation skip variable

* Fix translation warning on new subtitle copy

* Remove ticket creation prevention from 'Let us guide you' link