Add PoweredBy component (#103029)
This new component displays "Powered by" and then an Automattic brand
logo. Currently only wpcom, Woo and Jetpack logos are supported but
more can be added as necessary.

Logos are displayed at a uniform size and can be displayed in their
original colour, black or white.

It also brings support for these colour to JetpackLogo and WooLogo
components if they're being used directly.