Theme showcase: Enable preview for active theme (#104031)
The theme showcase allows each theme to be demoed by showing the demo
site in an iframe. However the active theme doesn't show the button to
open this overlaid iframe.

Seeing the theme demo site, and being able to click through and explore
it can still be useful for the active theme. This is doubly so for block
themes where the active theme on the site might have been substantially
modified.

This also brings `client/my-sites/theme/main.jsx` into compliance with the new react/display-name eslint rule.