Purchases: Update DataViews version to implement design feedback, third pass (#104118)
* Disable making "Type" and "Expiring soon" fields visible

These are both just here for the filtering. If we want to make the
columns visible, we could consider that for a future improvement.

* Pluralize "Plans" type filter to match "Domains"

* Allow selecting multiple sites in site filter at once

* Change status column header to be more clear

* Fix previousRoute for add/change payment method routes

It should only use previousRoute if it's a payment method list.

* Fix goToAddPaymentMethod in PurchaseItemPaymentMethod

It was using "add-payment-method" as the site slug and its arguments
were poorly named. This is a bug from
https://github.com/Automattic/wp-calypso/pull/103451