Add WP Admin util to open support links in Help Center (#104172)
Updates the Help Center app for WP Admin so any support link with a `data-target="wpcom-help-center"` attribute is opened in the Help Center directly rather than in a new tab.

This is useful for cases where React is not available and therefore we cannot use the `WpcomSupportLink` component introduced in Automattic/jetpack#43883.