Backup: fix finished screen on granular restores (#104151)
* Backup: move 'isFinished' to the state

When doing a granular restore, we cannot rely on the
'inProgressRewindStatus' value alone to decide whether to show the
finished/success page, given it might be storing stale information from
a previous run. Ensuring the 'isFinished' value is reset on every
restore and only using 'inProgressRewindStatus' to change its value at
the actual execution end is a safer approach.

* Fix progress bar not resetting

* Handle percent values in the state directly

This ensures we can reset its value when a new granular restore is
triggered, instead of showing the stale "100%" from the previous run.