Fix Help Center when notifications are not visible (#104389)
For some reason, the notifications button is not visible in WP Admin on some Atomic sites. This causes the Help Center not to load because there is some part of the logic that relies on that button.

To prevent that error, we're now ensuring that the notifications button exist before running the defendant logic.