Billing History: Preserve and restore view state in URL (#104538)
* Export constant for defaultSortDirection

* Preserve billing-history view properties in URL

* Use previousRoute for user-level receipt page back button

This will allow preserving the query string and thus the view options
like pagination from the list of receipts when the user returns to the
list from a receipt.

* Update URL keys to be unique to the DataView

This prevents calypso from re-using the keys from one DataView in
another when navigating between two (eg: from Billing History to Active
Upgrades).

This also will allow us to have the URL preserve the view for multiple
DataViews on the same page if we ever get around to implementing it for
the Memberships table which is on the same page as the purchases table.

* Use previousRoute for site-level receipt page back button