Jetpack Search purchase flow: Allow 'enter' key form submission during purchase flow (#104885)
* Allow 'enter' key form submission

This PR addresses the UX issue raised in #54344 by making it possible to use the 'enter' key to
submit the form at `/purchase-product/jetpack_search/monthly`. Because `components/suggestion-search/index.jsx`
was modified, this change also impacts any form that uses this component, meaning that 'enter' key
submission would not be limited to only the `/purchase-product/jetpack_search/monthly` flow. Attempts to
limit the context of this change proved more involved than anticipated, and if 'enter' key submission is
a UX concern for the Jetpack Search purchase flow, then I would expect it to be a concern for similar flows
elsewhere and so limiting the change is probably not warranted.

* Use <form> best standards

Use <form> for best standards

* revert unnecessary changes

---------

Co-authored-by: Jasper Kang <jasper.kang@automattic.com>