Revert "Hosting Dashboard: Link plan card to billing section (#105670)" (#105900)
The code being reverted worked fine in the v2 hosting dashboard, however
it didn't work fine in the v1 dashboard.

The v2 hosting dashboard has been retrofitted to the v1 sites overview,
but the billing section it links to is limited to the v2 hosting
dashboard and so the link does nothing in the v1 sites overview.

This reverts commit a52f52101effd21761ca75070d36720a1062d57d which means
the v1 and v2 hosting dashboards both link to the v1 billing section again.