Hosting Dashboard: Link plan card to billing section (#105922)
The Plan Card in the site overview now links to the new 'Plans &
Add-ons' billing section if the site has a paid plan and the v2
dashboard is being used.

Free sites, and the v1 dashboard continue to link to /plans in a new tab
as before.

This is a re-do of #105670 which was previously reverted by #105900.