Calypso Cleanup: Comments (#106132)
We forced everyone to the wp-admin Comments section in early 2025. The Calypso Comments section has been unused since then, largely containing only dead code.

This commit deletes that dead code, leaving only the section's router, to ensure that Calypso links are redirected to the wp-admin equivalent.

This commit does **not** remove the comments data layer and Redux state, as it's shared with the Reader.