Update the ai-site-builder flow to create a garden site (#105939)
* Allow the create-site step to create garden sites

For the PoC, we've added a query string parameter to the ai-site-builder flow. If this query param is present, we pass a garden_name and garden_partner_name to /sites/new so it creates a garden site on WP Cloud.

There are generic methods for setting garden_name and garden_partner name in the ONBOARD_STORE so those can be set at any time prior to the create-site step

* Improve garden site creation error message

* Don't add the big-sky-free-trial sticker to garden sites