Switch to the new version of the /all-domains endpoint for all domains lists in the Hosting Dashboard (#106475)
We want to introduce a new endpoint that would return the domain status alongside all the data that's necessary to show al list of domains. We'll use the same data to show all domains or site domains. It's a big change from our current approach where we had separate types for list of domains in the all domains view, a type for the site domains view and then the individual domain type.