Garden: CIAB hosting dashboard updates (#106489)
* Fix use-your-own-domain for Garden sites

When connecting an existing domain to a garden site, skip the plans step and checkout. Simply map the domain and continue.

* Fix so we can set a domain as primary

* Switch to doing site lookups by id instead of using find

* Preload sites on the domains route

* Memoize our sites by blog id

* Fix so we pass an array of sites on sites/domains

* Revert "Fix use-your-own-domain for Garden sites"

This reverts commit 1c17b7f436d360ebf79556cb5dbd467185b22283.

The redirect needs to happen in the submit method of the domains flow
instead.

* Allow garden sites to a map a domain from /ciab

For a garden site, bypass the checkout step when mapping a domain.
Simply map the domain and return to the site domain list