Increase clarity around retired themes (#107095)
On the theme details screen:
 
1.  Don't show a theme badge for retired themes
unless it is the users active theme. This reduces confusion where users
may think that upgrading plan will give them access to the retired
theme, while continuing to be clear that where it is the active theme,
that it's use may be contingent on their current plan.

2. Add a retired notice to the theme sheet this explains that the
theme is retired and what this means, with a link to the support centre
to learn more.