Garden: Domain mapping plan check (#106819)
* Revert "Allow garden sites to a map a domain from /ciab"

This reverts commit 5aaa8d12f308435de5451c195e8c4104c7fe1d91.

We'll be changing from a garden-specific check to a plan feature check.

* Add domain mapping feature to feature-list

* Tweak hasPlanFeature to accept either a Site or SiteDetails

* Update post-domain mapping redirect

For CIAB sites, we map the domain and redirect without checking out if
they already have a plan.

Previously, we had an isGarden check which redirected back to the /ciab
dashboard. Now we use a redirect_to query parameter.

* Update SiteDomains redirect to use the current url