Domain Transfer Redesign: Make small updates (#107540)
* Do not show transfer failure page if last_transfer_error is an empty string

* Make auth code input field password masked

* Redirect user to site domain list after providing auth code

Previously, the user was redirected to the transfer progress page.
However, the user saw the initial transfer step, as if the transfer
hadn't started yet, since the domain status wasn't refreshed with the
latest data yet.

* Invalidate the domain query when the transfer starts successfully

After the transfer begins, we have to fetch the domain data from the
backend again so the domain transfer status is updated.

* Only show cancel link in the transfer setup page if `purchase` is defined

* Add "Cancel transfer" link to the transfer progress page

* Do not show "Cancel transfer" link when the transfer is in progress